nilfs2: fix sanity check of btree level in nilfs_btree_root_broken()
The range check for b-tree level parameter in nilfs_btree_root_broken() is
wrong; it accepts the case of "level == NILFS_BTREE_LEVEL_MAX" even though
the level is limited to values in the range of 0 to (NILFS_BTREE_LEVEL_MAX
- 1).

Since the level parameter is read from storage device and used to index
nilfs_btree_path array whose element count is NILFS_BTREE_LEVEL_MAX, it
can cause memory overrun during btree operations if the boundary value is
set to the level parameter on device.

This fixes the broken sanity check and adds a comment to clarify that the
upper bound NILFS_BTREE_LEVEL_MAX is exclusive.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>