include/linux/wait.h: change wait_on_bit*() to take an unsigned long*, not a void*
The implementations of wait_on_bit*() will only work with long-aligned
memory on systems that don't support misaligned loads and stores.  This
patch changes the function prototypes to ensure that the compiler will
enforce alignment.

Running

  make defconfig
  make KFLAGS="-Werror"

seems to indicate that, as of c56fb6564dcd ("Fix a misaligned load inside
ptrace_attach()"), there are now no users of non-long-aligned calls to
wait_on_bit*().  I additionally tried a few "make randconfig" attempts,
none of which failed to compile for this reason.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>
Reviewed-by: Chris Metcalf <cmetcalf@ezchip.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Richard Weinberger <richard@nod.at>
Cc: "Paul E. McKenney" <paulmck@linux.vnet.ibm.com>
Cc: Pranith Kumar <bobby.prani@gmail.com>
Cc: Vladimir Davydov <vdavydov@parallels.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>