zpool: remove zpool_evict()
Remove zpool_evict() helper function.  As zbud is currently the only zpool
implementation that supports eviction, add zpool and zpool_ops references
to struct zbud_pool and directly call zpool_ops->evict(zpool, handle) on
eviction.

Currently zpool provides the zpool_evict helper which locks the zpool list
lock and searches through all pools to find the specific one matching the
caller, and call the corresponding zpool_ops->evict function.  However,
this is unnecessary, as the zbud pool can simply keep a reference to the
zpool that created it, as well as the zpool_ops, and directly call the
zpool_ops->evict function, when it needs to evict a page.  This avoids a
spinlock and list search in zpool for each eviction.

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Cc: Seth Jennings <sjennings@variantweb.net>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Nitin Gupta <ngupta@vflare.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>